move model methods to BarracudaModelExtensions
